requirements.txt
# base
chardet==5.1.0
filetype==1.2.0
python-magic==0.4.27
nltk==3.8.1
tabulate==0.9.0
requests==2.31.0
urllib3==1.26.16
beautifulsoup4==4.12.2
emoji==2.8.0

# doc and docx
lxml==4.9.3
python-docx==0.8.11

# csv, tsv
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3

# epub
ebooklib==0.18

# markdown
importlib-metadata==6.8.0
markdown==3.4.4
zipp==3.16.2

# msg
msg-parser==1.2.0
olefile==0.46

# odt, pandoc, rtf, rst, org
pypandoc==1.11

# pdf
pdf2image==1.16.3
pdfminer-six==20221105
pdfplumber==0.10.2
wheel==0.41.0
pypdfium2==4.18.0
PyMuPDF==1.23.2
opencv-python==4.8.0.76
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
fonttools==4.42.1
idna==3.4
scipy==1.10.1
shapely==2.0.1
pydantic==1.10.12

# pptx
pillow==10.0.0
python-pptx==0.6.21
xlsxwriter==3.1.2

# xlsx
et-xmlfile==1.1.0
openpyxl==3.1.2
xlrd==2.0.1